fixing some issues, revertig change to Doc back to Map<String,Object>. Making BasicFunctionalityTest work (partly)

+++ src/de/caluga/morphium/driver/commands/CountMongoCommand.java
@@ -97,8 +97,8 @@ public class CountMongoCommand extends MongoCommand<CountMongoCommand> implement
         return getSingleResultFor(id);
     }
 
-    public long getCount() throws MorphiumDriverException {
-        return (long) execute().get("n");
+    public int getCount() throws MorphiumDriverException {
+        return (int) execute().get("n");
     }
